requirements.txt
python-telegram-bot==20.6
fastapi
uvicorn
yt-dlp